+++ viz/tap/Dockerfile
@@ -13,9 +13,11 @@ RUN ./bin/install-deps $TARGETARCH
 FROM go-deps as golang
 WORKDIR /linkerd-build
 COPY pkg pkg
+# TODO: remove after https://github.com/linkerd/linkerd2/issues/5661
 COPY controller controller
+# TODO: remove when BuildResource is refactored
+# https://github.com/linkerd/linkerd2/issues/5589
 COPY viz/metrics-api/gen/viz viz/metrics-api/gen/viz
-COPY viz/metrics-api/client viz/metrics-api/client
 COPY viz/tap viz/tap
 COPY viz/pkg viz/pkg
