+++ viz/tap/api/grpc_server.go
@@ -1,4 +1,4 @@
-package controller
+package api
 
 import (
 	"context"
