+++ viz/tap/cmd/main.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/linkerd/linkerd2/viz/tap/controller"
+	"github.com/linkerd/linkerd2/viz/tap/api"
 	"github.com/linkerd/linkerd2/viz/tap/injector"
 )
 
@@ -14,8 +14,8 @@ func main() {
 		os.Exit(1)
 	}
 	switch os.Args[1] {
-	case "controller":
-		controller.Main(os.Args[2:])
+	case "api":
+		api.Main(os.Args[2:])
 	case "injector":
 		injector.Main(os.Args[2:])
 	}
